Replace TRS inside processor_2015
replaces the transform rule system class inside processor_2015 by inlining the relevant code from TRS into the processor algorithm. This does not remove the now redundant TRS code, which is left to a seperate commit since it may still be used elsewhere in the codebase and previous attempts to rip and replace have resulted in a thousand wilting branches in my local repo. .patch files litter the directory, configuration gets clogged on its way to child objects, and the whole tree is threatened if we don't do some army-of-darkness style civil war surgery.